feat: added XML docs to the example app and restructured the project
